middle_rate/requirements.txt
kafka-python
requests
selenium
bs4
pandas
pytz